Ports: Do not download sources if they're already present
When running ./package.sh to rebuild an already installed port, we would not
want to spend time re-downlodaing the same tarball again. Ideally, this should
use some sort of hash checking to ensure the file is not truncated or something,
but this is good enough for now.

+++ Ports/gcc/package.sh
@@ -9,8 +9,12 @@ installopts="DESTDIR=$SERENITY_ROOT/Root install-gcc install-target-libgcc insta
 depends="binutils"
 
 fetch() {
-    read url filename <<< $(echo "$files")    
-    run_nocd curl -O "$url" -o "$filename"
+    read url filename <<< $(echo "$files")
+    if [ -f "$filename" ]; then
+        echo "$filename already exists"
+    else
+        run_nocd curl -O "$url" -o "$filename"
+    fi
     run_nocd tar xf "$filename"
     run contrib/download_prerequisites
     for f in patches/*; do
